feat: add `normalize_saturating()`
That way, one can express another way with which Git handles worktrees.
