+++ gix-path/tests/path/convert/mod.rs
@@ -20,6 +20,91 @@ fn assure_windows_separators() {
 
 mod normalize;
 
+mod normalize_and_clean {
+    use std::{borrow::Cow, path::Path};
+
+    use gix_path::normalize_and_clean;
+
+    fn p(input: &str) -> &Path {
+        Path::new(input)
+    }
+
+    #[test]
+    fn clean_paths_and_use_the_cwd_for_empty_results() {
+        let cwd = p("/cwd");
+        for (input, expected) in [
+            ("", "/cwd"),
+            (".", "/cwd"),
+            ("./a", "a"),
+            ("a/./b", "a/b"),
+            ("a//b/", "a/b"),
+            ("a/..", "/cwd"),
+        ] {
+            assert_eq!(
+                normalize_and_clean(p(input).into(), cwd).expect("path can be normalized"),
+                p(expected),
+                "'{input}' cleans to '{expected}'"
+            );
+        }
+        assert_eq!(
+            normalize_and_clean(p(".").into(), p(""))
+                .expect("path can be normalized")
+                .as_ref(),
+            p(""),
+            "an empty CWD allows an empty normalized path"
+        );
+        assert_eq!(
+            normalize_and_clean(p(".").into(), cwd)
+                .expect("path can be normalized")
+                .as_ref(),
+            cwd,
+            "an empty input path is the CWD"
+        );
+        assert!(
+            matches!(normalize_and_clean(p("a/b").into(), cwd), Some(Cow::Borrowed(path)) if path == p("a/b")),
+            "already-clean borrowed paths stay borrowed"
+        );
+    }
+}
+
+mod normalize_saturating {
+    use std::{borrow::Cow, path::Path};
+
+    use gix_path::normalize_saturating;
+
+    fn p(input: &str) -> &Path {
+        Path::new(input)
+    }
+
+    #[test]
+    fn walking_up_too_much_stays_at_the_root() {
+        let cwd = "/users/name".as_ref();
+        assert_eq!(
+            normalize_saturating(p("./a/b/../../../../../actually-valid").into(), cwd).as_ref(),
+            p("/actually-valid")
+        );
+        assert_eq!(
+            normalize_saturating(p("/a/b/../../../../actually-valid").into(), cwd).as_ref(),
+            p("/actually-valid")
+        );
+        assert_eq!(
+            normalize_saturating(p("/a/b/../../../../..").into(), cwd).as_ref(),
+            p("/")
+        );
+    }
+
+    #[test]
+    fn preserves_normal_paths() {
+        let path = p("a/b");
+        assert!(matches!(normalize_saturating(path.into(), p("/users")), Cow::Borrowed(actual) if actual == path));
+        assert_eq!(
+            normalize_saturating(p("a/../b").into(), p("/users")).as_ref(),
+            p("b"),
+            "paths that do not reach the root normalize as usual"
+        );
+    }
+}
+
 mod join_bstr_unix_pathsep {
     use bstr::BStr;
     use gix_path::join_bstr_unix_pathsep;
